backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.32.0,<3.0.0
gitpython>=3.1.43,<4.0.0
crewai==1.6.1
litellm==1.81.13
tavily-python>=0.5.0,<1.0.0
PyGithub>=2.3.0,<3.0.0
pymongo>=4.7.2,<5.0.0
redis>=5.0.8,<6.0.0
